E2C: Resources table refactor
