+++ public/app/features/migrate-to-cloud/api/index.ts
@@ -1,4 +1,7 @@
 export * from './endpoints.gen';
+import { BaseQueryFn, QueryDefinition } from '@reduxjs/toolkit/dist/query';
+import { A } from 'msw/lib/core/HttpResponse-vQNlixkj';
+
 import { generatedAPI } from './endpoints.gen';
 
 export const cloudMigrationAPI = generatedAPI.enhanceEndpoints({
@@ -35,5 +38,24 @@ export const cloudMigrationAPI = generatedAPI.enhanceEndpoints({
     runCloudMigration: {
       invalidatesTags: ['cloud-migration-run'],
     },
+
+    getDashboardByUid(endpoint) {
+      suppressErrorsOnQuery(endpoint);
+    },
   },
 });
+
+function suppressErrorsOnQuery<QueryArg, BaseQuery extends BaseQueryFn, TagTypes extends string, ResultType>(
+  endpoint: QueryDefinition<QueryArg, BaseQuery, TagTypes, ResultType>
+) {
+  if (!endpoint.query) {
+    return;
+  }
+
+  const originalQuery = endpoint.query;
+  endpoint.query = (...args) => {
+    const baseQuery = originalQuery(...args);
+    baseQuery.showErrorAlert = false;
+    return baseQuery;
+  };
+}
